// src/common/constants/exception-code.enum.ts
export enum ExceptionCode {
  ACCESS_TOKEN_MISSING = 'ACCESS_TOKEN_MISSING',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USERNAME_ALREADY_USED = 'USERNAME_ALREADY_USED',
  EMAIL_ALREADY_USED = 'EMAIL_ALREADY_USED',
  EMAIL_TOKEN_INVALID = 'EMAIL_TOKEN_INVALID',
  EMAIL_SEND_FAILED = 'EMAIL_SEND_FAILED',
  ALREADY_LINKED = 'ALREADY_LINKED',
  NO_REMAINING_AUTH = 'NO_REMAINING_AUTH',
  REFRESH_TOKEN_NOT_FOUND = 'REFRESH_TOKEN_NOT_FOUND',
  REFRESH_TOKEN_MISMATCH = 'REFRESH_TOKEN_MISMATCH',
  REFRESH_TOKEN_INVALID = 'REFRESH_TOKEN_INVALID',
  REFRESH_TOKEN_SAVE_FAILED = 'REFRESH_TOKEN_SAVE_FAILED',
  REFRESH_TOKEN_MISSING = 'REFRESH_TOKEN_MISSING',
  ACCESS_TOKEN_BLACKLISTED = 'ACCESS_TOKEN_BLACKLISTED',
  ACCESS_TOKEN_INVALID = 'ACCESS_TOKEN_INVALID',
  NO_LINKED_ACCOUNT = 'NO_LINKED_ACCOUNT',
  OAUTH_TOKEN_GENERATION_FAILED = 'OAUTH_TOKEN_GENERATION_FAILED',
  SAME_PASSWORD_USED = 'SAME_PASSWORD_USED',
  PASSWORD_REUSED = 'PASSWORD_REUSED',
  ALREADY_HAS_LOCAL_ACCOUNT = 'ALREADY_HAS_LOCAL_ACCOUNT',
  TEST_ACCOUNT_MUTATION = 'TEST_ACCOUNT_MUTATION',
}